pyproject.toml
[project]
name = "edge"
version = "v0.0.1"
description = ""
authors = [
    {name = "ghostminer", email = "1696415745@qq.com"},
]
dependencies = [
    "logging>=0.4.9.6",
    "tornado>=6.4",
    "mysql-connector-python>=8.2.0",
    "toml>=0.10.2",
    "uuid>=1.30",
    "asyncio>=3.4.3",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
start = "src/__main__.py"
test = "tests/__main__.py"
